+++ sdk/src/main/java/com/eyeson/sdk/events/PresentationResponse.kt
@@ -1,16 +1,17 @@
 package com.eyeson.sdk.events
 
-enum class MediaPlaybackResponse(val responseCode: Int) {
+enum class PresentationResponse(val responseCode: Int) {
     UNSPECIFIED(-7),
     OK(200),
     CREATED(201),
     BAD_REQUEST(400),
     NOT_FOUND(404),
+    CONFLICT(409),
     GONE(410),
     ERROR(500);
 
     companion object {
-        private val mapping = entries.associateBy(MediaPlaybackResponse::responseCode)
+        private val mapping = entries.associateBy(PresentationResponse::responseCode)
         fun fromResponseCode(code: Int) = mapping[code] ?: UNSPECIFIED
         fun isSuccessful(code: Int): Boolean {
             return code == OK.responseCode || code == CREATED.responseCode
@@ -18,6 +19,6 @@ enum class MediaPlaybackResponse(val responseCode: Int) {
     }
 
     fun isSuccessful(): Boolean {
-        return MediaPlaybackResponse.isSuccessful(responseCode)
+        return PresentationResponse.isSuccessful(responseCode)
     }
 }